feat(web): preload home page queries
